Create a new class, BookMethods, to handle books
